requirements.txt
azure-functions
sendgrid
PyNFe
